Move the specs in a separate folder
